A little more progress.  Fixed the warnings on startup about SLF4J's not being able to find a backing logger.  This was due to slf4j having changed the way it finds and loads logger implementations and the startup script not having adjusted for that - or maybe I at some point moved the slf4j dependency version up.  Regardless, moving up to 1.3.5 and adding reload4j into the dependency mix fixed the warnings.  As an added bonus, my thing little classes started generating log output for me.
The other big win was figuring out why the loader servlet was not actually responding to get requests, despite a get action having been added.  Turns out, scalatra
relies on conventions and hidden magic.  The opening curly brace of the action definition must be on the same line as the HTTP method declaration: that is,

get("/route") {

and not

get("/route")
{

Being old and not particularly interested in following Java conventions in this project I fell back into what I was taught which does not work.
This does not appear to be documented anywhere, unfortunately.